refactor: Use newer version of `NewBarChart` for all bar-charts
