+++ src/components/Base/Charts/index.ts
@@ -1,3 +1,2 @@
 export * from './PieChart.component';
 export * from './BarChart.component';
-export * from './NewBarChart.component';
